requirements.txt
torch>=1.12.0
torchvision
numpy
pyyaml
scikit-image
opencv-python
tensorboard
torch-scatter
torch-sparse
torch-geometric
